feat: support absolute path
The github repo support relative path and the links starting with `/` will be
relative to the repository root. It's easy to link to files, such as
MAINTAINERS, LICENSE and so on.

This patch is to introduce `allow-absolute-patch` flag to allow absolute
path to join with `--root` and evaluate the file link.

REF:

* https://github.blog/2013-01-31-relative-links-in-markup-files/
* https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#relative-links

Signed-off-by: Wei Fu <fuweid89@gmail.com>